requirements.txt
mysql-connector-python
gensim
spacy
sumy
ginza
summarization
mecab-python3
unidic-lite
nltk
ja-ginza
